+++ src/lib.rs
@@ -228,3 +228,6 @@ pub use redox::redoxable::*;
 
 #[allow(unused_imports)]
 pub use util::low_level::*;
+
+#[allow(unused_imports)]
+pub use structure::dataframe::*;