+++ src/structure/mod.rs
@@ -13,3 +13,4 @@ pub mod matrix;
 pub mod multinomial;
 pub mod polynomial;
 pub mod vector;
+pub mod dataframe;